Add further styling and styles
Add further universal styling for the DoughFlow Dashboard. Add functionality for deleting categories by id for that specific budget.

+++ client/src/views/BudgetManagement.vue
@@ -1,17 +1,18 @@
 <template>
   <div class="everything">
-    <Navbar id="123"/>
+    <Navbar id="navBar"/>
 
     <div class="budget-management container-fluid">
       <div class="row">
 
         <!-- Left Menu: List of Budgets -->
         <div class="left-menu col-md-3">
-          <h2>Budgets</h2>
+          <h2 id="headerText" class="white-text">Budgets</h2>
           <ul class="list-unstyled">
             <li v-for="budget in budgets" :key="budget._id">
               <div class="budget-box p-2 mb-3" @click.stop="handleBudgetClick(budget._id)">
                 {{ budget.name }} - ${{ budget.amount }}
+                <button class="btn btn-info btn-sm ml-2" @click.stop="selectBudget(budget)">Select Budget</button>
                 <button class="btn btn-danger btn-sm ml-2" @click.stop="deleteBudget(budget._id)">Delete</button>
               </div>
             </li>
@@ -23,21 +24,26 @@
 
           <!-- Right Section: Create Budget Form -->
           <div class="right-section mb-4">
-            <h2>Create New Budget</h2>
+            <h2 id="headerText" class="white-text">Create New Budget</h2>
             <form @submit.prevent="createBudget" class="form-inline">
-              <input v-model="newBudget.name" placeholder="Budget Name" required class="form-control mr-2">
-              <input v-model.number="newBudget.amount" placeholder="Amount" required class="form-control mr-2">
+              <div class="create-budget-fields">
+              <input v-model="newBudget.name" placeholder="Budget Name" required class="form-control mr-2" maxlength="10">
+              <input v-model.number="newBudget.amount" placeholder="Amount" required class="form-control mr-2" maxlength="8">
               <button type="submit" class="btn btn-primary">Add Budget</button>
+            </div>
             </form>
           </div>
 
           <div class="category-section" v-if="selectedBudget">
             <div v-if="categories && categories.length">
-              <h2>Categories for {{ selectedBudget.name }}</h2>
+              <h2 class="white-text">Categories for {{ selectedBudget.name }}</h2>
               <ul class="list-unstyled">
                 <li v-for="category in categories" :key="category._id" class="p-2 mb-3 border rounded">
-                  <div class="category-box">
-                    <strong>{{ category.categoryName }}</strong>
+                  <div class="category-container">
+                    <div class="category-box">
+                      <div class="category-name">
+                        <strong>{{ category.categoryName }}</strong>
+                      </div>
                     <div>Total amount spent: {{ sumExpenses(category.expenses) }}kr</div>
                     <ul class="list-unstyled">
                       <li v-for="expense in category.expenses" :key="expense._id" class="mt-2">
@@ -46,11 +52,13 @@
                       </li>
                     </ul>
                   </div>
+                  <button class="btn btn-danger btn-sm ml-2" @click.stop="deleteCategory(category._id)">Delete Category</button>
+                </div>
                 </li>
               </ul>
             </div>
             <div v-else>
-              <h2>No categories for {{ selectedBudget.name }}</h2>
+              <h2 class="white-text">No categories for {{ selectedBudget.name }}</h2>
             </div>
           </div>
         </div>
@@ -75,14 +83,15 @@ export default {
       budgets: [],
       selectedBudget: null,
       categories: [],
-      expenses: []
+      expenses: [],
+      URL: 'http://localhost:3000/api/v1'
     }
   },
   components: {
     Navbar
   },
   mounted() {
-    axios.get('http://localhost:3000/api/v1/budgets')
+    axios.get(`${this.URL}/budgets`)
       .then(response => {
         if (Array.isArray(response.data)) {
           this.budgets = response.data
@@ -95,9 +104,10 @@ export default {
       })
   },
   methods: {
+
     async createBudget() {
       try {
-        const response = await fetch('http://localhost:3000/api/v1/budgets', {
+        const response = await fetch(`${this.URL}/budgets`, {
           method: 'POST',
           headers: {
             'Content-Type': 'application/json'
@@ -116,41 +126,52 @@ export default {
         console.error('Error creating budget:', error)
       }
     },
+
+    selectBudget(budget) {
+      this.selectedBudget = budget
+      alert(`${this.selectedBudget.name} is now set as your budget.`)
+    },
+
+    async fetchCategoriesForBudget(budgetId) {
+      const response = await fetch(`${this.URL}/budgets/${budgetId}/categories`)
+      if (!response.ok) {
+        throw new Error('Error fetching categories')
+      }
+      return response.json()
+    },
+
+    async fetchExpensesForCategory(categoryId) {
+      const response = await fetch(`${this.URL}/categories/${categoryId}/expenses`)
+      if (!response.ok) {
+        throw new Error('Error fetching expenses for category')
+      }
+      return response.json()
+    },
+
     async handleBudgetClick(budgetId) {
       this.categories = []
       try {
         const budget = this.budgets.find(b => b._id === budgetId)
         this.selectedBudget = budget
 
-        const response = await fetch(`http://localhost:3000/api/v1/budgets/${budgetId}/categories`)
-
-        if (!response.ok) {
-          const data = await response.json()
-          console.error('Error fetching categories:', data.error)
-          return
-        }
-
-        const responseData = await response.json()
-        console.log('Fetched Categories:', responseData)
-        // Update the local categories array with fetched data
-        this.categories = await Promise.all(
-          responseData.categories.map(async category => {
-            const expensesResponse = await fetch(`http://localhost:3000/api/v1/categories/${category._id}/expenses`)
-            const expensesData = await expensesResponse.json()
-            category.expenses = (expensesData && expensesData.expenses) || []
-            return category
-          })
-        )
-
-        this.$forceUpdate()
+        const categoryData = await this.fetchCategoriesForBudget(budgetId)
+        this.categories = await Promise.all(categoryData.categories.map(async category => {
+          const expensesData = await this.fetchExpensesForCategory(category._id)
+          category.expenses = expensesData.expenses || []
+          return category
+        }))
       } catch (error) {
-        console.error('Error fetching categories:', error)
+        console.error('Error:', error.message)
       }
     },
 
     async deleteBudget(id) {
+      const confirmed = window.confirm('Are you sure you want to delete this budget? This action cannot be undone!')
+      if (!confirmed) {
+        return
+      }
       try {
-        const response = await fetch(`http://localhost:3000/api/v1/budgets/${id}`, {
+        const response = await fetch(`${this.URL}/budgets/${id}`, {
           method: 'DELETE'
         })
         if (response.ok) {
@@ -170,7 +191,7 @@ export default {
         return
       }
       try {
-        const response = await fetch('http://localhost:3000/api/v1/budgets/', {
+        const response = await fetch(`${this.URL}/budgets/`, {
           method: 'DELETE'
         })
         if (!response.ok) {
@@ -186,9 +207,33 @@ export default {
         console.error('Error deleting all budgets:', error)
       }
     },
+
+    async deleteCategory(categoryId) {
+      const confirmed = window.confirm('Are you sure you want to delete this category? This action cannot be undone!')
+      if (!confirmed) {
+        return
+      }
+
+      try {
+        const response = await fetch(`${this.URL}/categories/${categoryId}`, {
+          method: 'DELETE'
+        })
+
+        if (response.ok) {
+          // Remove the deleted category from the categories list
+          this.categories = this.categories.filter(category => category._id !== categoryId)
+        } else {
+          const data = await response.json()
+          alert(data.error)
+        }
+      } catch (error) {
+        console.error('Error deleting category:', error)
+      }
+    },
+
     async fetchExpensesByCategory(categoryId) {
       try {
-        const response = await axios.get(`http://localhost:3000/api/v1/categories/${categoryId}/expenses`)
+        const response = await axios.get(`${this.URL}/categories/${categoryId}/expenses`)
         if (response.data && response.data.expenses && response.data.expenses.length > 0) {
           this.expenses = response.data.expenses
         } else {
@@ -234,41 +279,45 @@ export default {
   justify-content: space;
   min-width: 20px;
   background:#1f8cdb;
-  border-right: 3px solid #ffffff;
-  padding-right: 10px;
-  margin-right: 10px;
 }
 
 .left-menu li {
   margin-bottom: 10px;
 }
+.left-menu ul li:last-child {
+    margin-bottom: 20px;
+}
+.left-menu ul {
+  min-height: calc(100%-100px);
+}
 
 .right-section {
   flex: 3;
+  flex-wrap: nowrap;
 }
 
 .right-content {
   display: flex;
   flex: 2;
   flex-direction: column;
+  flex-wrap: nowrap;
   background: #7fc9ff;
 }
 .budget-box {
   display: inline-block;
   padding: 5px 10px;
   border: 1px solid #42bbf7;
-  margin-right: 10px; /* To give space between the box and the delete button */
+  background: #ffffff;
+  margin-right: 10px;
   cursor: pointer;
   border-radius: 4px;
   transition: background-color 0.3s;
-
-  /* Optional: add hover effect for better UX */
   &:hover {
-    background-color: #e5eff0;
+    background-color: #57acbb;
   }
 }
 
-/* Remove bullet point from list items */
+/* Remove bullet point dots from list*/
 .left-menu ul {
   list-style-type: none;
   /*padding-left: 0;*/
@@ -285,6 +334,7 @@ export default {
 .category-section li {
   padding: 5px 10px;
   border: 1px solid #42bbf7;
+  background: #ffffff;
   margin-bottom: 10px;
   border-radius: 4px;
   transition: background-color 0.3s;
@@ -299,9 +349,9 @@ export default {
 .category-box {
     display: inline-block;
     padding: 5px 10px;
-    border: 1px solid #42bbf7;
-    margin-right: 10px; /* To give space between boxes */
-    margin-bottom: 10px; /* To give space between rows of boxes */
+    border: 1px solid #ffffff;
+    margin-right: 10px;
+    margin-bottom: 10px;
     cursor: pointer;
     border-radius: 4px;
     transition: background-color 0.3s;
@@ -312,12 +362,21 @@ export default {
         background-color: #53b1c1;
     }
 }
+.category-container {
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  align-items: center;
+}
+.delete-category-button {
+  margin-top: 10px;
+}
 .delete-all-button {
     position: absolute;
     bottom: 10px;
     left: 50%;
     transform: translateX(-50%);
-    margin-top: 20px;
+    margin-top: 10%;
     background-color: #ff4d4d;
     color: white;
     border: none;
@@ -329,10 +388,30 @@ export default {
     &:hover {
         background-color: #b71a1a;
     }
-
 }
-#123 {
+.create-budget-fields {
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  align-items: center;
+  width: 100%;
+}
+.form-control, .btn-primary {
+    margin: 5px;
+}
+.form-control {
+  min-width: 150px;
+}
+.white-text {
+  color:#ffffff;
+}
+.category-name {
+  font-size: 24px;
+}
+#navBar {
   align-items: flex-start;
 }
-
+#headerText {
+  margin-top: 15px;
+}
 </style>
